feat(transformer): use train-val split instead of test set for val
